Renamed Task class to Consts, moved classes from task 3 (hw2) to separate files
