+++ src/main/java/edu/Consts.java
@@ -3,9 +3,12 @@ package edu;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class Task {
+public class Consts {
     protected static final Logger LOGGER = LogManager.getLogger();
     protected static final int DECIMAL_BASE = 10;
+    protected static final int MAX_PROBABILITY = 100;
+    protected static final int HIGH_PROBABILITY = 80;
+    protected static final int EVEN_PROBABILITY = 50;
 
-    protected Task() {}
+    protected Consts() {}
 }
