+++ src/main/java/edu/hw1/Task2.java
@@ -1,8 +1,8 @@
 package edu.hw1;
 
-import edu.Task;
+import edu.Consts;
 
-public final class Task2 extends Task {
+public final class Task2 extends Consts {
     private Task2() {
     }
