+++ src/main/java/edu/hw2/Task4.java
@@ -1,12 +1,16 @@
 package edu.hw2;
 
-import edu.Task;
+import edu.Consts;
 
-public final class Task4 extends Task {
-    public static CallingInfo callingInfo() {
-        StackTraceElement stackTrace = new Throwable().getStackTrace()[1];
-        return new CallingInfo(stackTrace.getClassName(), stackTrace.getMethodName());
+public final class Task4 extends Consts {
+    private Task4() {
     }
 
-    public record CallingInfo(String className, String methodName) {}
+    public static CallingInfo callingInfo() {
+        StackTraceElement stackTraceLastCall = new Throwable().getStackTrace()[1];
+        return new CallingInfo(stackTraceLastCall.getClassName(), stackTraceLastCall.getMethodName());
+    }
+
+    public record CallingInfo(String className, String methodName) {
+    }
 }
