+++ src/main/java/edu/hw2/task3/ConnectionException.java
@@ -1,0 +1,11 @@
+package edu.hw2.task3;
+
+public class ConnectionException extends RuntimeException {
+    public ConnectionException(String message) {
+        super(message);
+    }
+
+    public ConnectionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
