+++ src/main/java/edu/hw2/task3/connection/Connection.java
@@ -1,0 +1,5 @@
+package edu.hw2.task3.connection;
+
+public interface Connection extends AutoCloseable {
+    void execute(String command);
+}
