+++ src/main/java/edu/hw2/task3/connection_manager/ConnectionManager.java
@@ -1,0 +1,7 @@
+package edu.hw2.task3.connection_manager;
+
+import edu.hw2.task3.connection.Connection;
+
+public interface ConnectionManager {
+    Connection getConnection();
+}
