+++ src/main/java/edu/hw2/task3/connection_manager/FaultyConnectionManager.java
@@ -1,0 +1,12 @@
+package edu.hw2.task3.connection_manager;
+
+import edu.Consts;
+import edu.hw2.task3.connection.Connection;
+import edu.hw2.task3.connection.FaultyConnection;
+
+public class FaultyConnectionManager extends Consts implements ConnectionManager {
+    @Override
+    public Connection getConnection() {
+        return new FaultyConnection();
+    }
+}
